pyproject.toml
[project]
name = "ChatUSP"
version = "0.1.0"
description = "Chatbot da USP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.45.2",
    "crawl4ai>=0.4.247",
    "fastapi>=0.115.6",
    "instructor>=1.7.2",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "psycopg-binary>=3.2.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
]
